+++ packages/addons/sitemap/admin/components/List/Collection/Row.jsx
@@ -3,6 +3,7 @@ import React from 'react';
 import { Pencil, Trash } from '@strapi/icons';
 import { Box, Flex, Tr, Td, Typography, IconButton } from '@strapi/design-system';
 import { useSelector } from 'react-redux';
+import getSelectedContentType from '../../../helpers/getSelectedContentType';
 
 const CustomRow = ({ openModal, entry }) => {
   const contentTypes = useSelector((store) => store.getIn(['sitemap', 'contentTypes'], {}));
@@ -20,20 +21,16 @@ const CustomRow = ({ openModal, entry }) => {
   return (
     <Tr key={entry.id}>
       <Td>
-        <Typography variant="omega" textColor="neutral800">{contentTypes[entry.name] && contentTypes[entry.name].displayName}</Typography>
-      </Td>
-      <Td>
-        <Typography variant="omega" textColor="neutral800">{entry.langcode === 'und' ? 'N/A' : entry.langcode}</Typography>
-      </Td>
-      <Td>
-        <Typography variant="omega" textColor="neutral800">{entry.pattern}</Typography>
+        <Typography variant="omega" textColor="neutral800">{getSelectedContentType(contentTypes, entry.name) && getSelectedContentType(contentTypes, entry.name).name}</Typography>
       </Td>
       <Td>
         <Flex>
-          <IconButton onClick={handleEditClick} label="Edit" noBorder icon={<Pencil />} />
-          <Box paddingLeft={1}>
-            <IconButton onClick={handleDeleteClick} label="Delete" noBorder icon={<Trash />} />
-          </Box>
+          <Flex marginLeft="auto">
+            <IconButton onClick={handleEditClick} label="Edit" noBorder icon={<Pencil />} />
+            <Box paddingLeft={1}>
+              <IconButton onClick={handleDeleteClick} label="Delete" noBorder icon={<Trash />} />
+            </Box>
+          </Flex>
         </Flex>
       </Td>
     </Tr>