+++ packages/addons/sitemap/server/services/query.js
@@ -5,70 +5,11 @@ import xml2js from 'xml2js';
 
 const parser = new xml2js.Parser({ attrkey: 'ATTR' });
 
-import { noLimit, getService, logMessage } from '../utils';
+import { noLimit, logMessage } from '../utils';
 
 /**
  * Query service.
  */
-
-/**
- * Get an array of fields extracted from all the patterns across
- * the different languages.
- *
- * @param {obj} contentType - The content type
- * @param {bool} topLevel - Should include only top level fields
- * @param {bool} isLocalized - Should include the locale field
- * @param {string} relation - Specify a relation. If you do; the function will only return fields of that relation.
- *
- * @returns {array} The fields.
- */
-const getFieldsFromConfig = (contentType, topLevel = false, isLocalized = false, relation = null) => {
-  let fields = [];
-
-  if (contentType) {
-    Object.entries(contentType['languages']).map(([langcode, { pattern }]) => {
-      fields.push(...getService('pattern').getFieldsFromPattern(pattern, topLevel, relation));
-    });
-  }
-
-  if (topLevel) {
-    if (isLocalized) {
-      fields.push('locale');
-    }
-
-    fields.push('updatedAt');
-  }
-
-  // Remove duplicates
-  fields = [...new Set(fields)];
-
-  return fields;
-};
-
-/**
- * Get an object of relations extracted from all the patterns across
- * the different languages.
- *
- * @param {obj} contentType - The content type
- *
- * @returns {object} The relations.
- */
-const getRelationsFromConfig = (contentType) => {
-  const relationsObject = {};
-
-  if (contentType) {
-    Object.entries(contentType['languages']).map(([langcode, { pattern }]) => {
-      const relations = getService('pattern').getRelationsFromPattern(pattern);
-      relations.map((relation) => {
-        relationsObject[relation] = {
-          fields: getFieldsFromConfig(contentType, false, false, relation),
-        };
-      });
-    });
-  }
-
-  return relationsObject;
-};
 
 /**
  * Query the nessecary pages from Strapi to build the sitemap with.
@@ -82,9 +23,6 @@ const getRelationsFromConfig = (contentType) => {
 const getPages = async (config, contentType, ids) => {
   const excludeDrafts = config.excludeDrafts && strapi.contentTypes[contentType].options.draftAndPublish;
   const isLocalized = strapi.contentTypes[contentType].pluginOptions?.i18n?.localized;
-
-  const relations = getRelationsFromConfig(config.contentTypes[contentType]);
-  const fields = getFieldsFromConfig(config.contentTypes[contentType], true, isLocalized);
 
   const pages = await noLimit(strapi, contentType, {
     filters: {
@@ -105,13 +43,33 @@ const getPages = async (config, contentType, ids) => {
       } : {},
     },
     locale: 'all',
-    fields,
+    fields: isLocalized ? 'locale' : undefined,
     populate: {
-      localizations: {
-        fields,
-        populate: relations,
+      url_alias: {
+        populate: 'url_path',
       },
-      ...relations,
+      localizations: {
+        fields: 'locale',
+        filters: {
+          $or: [
+            {
+              sitemap_exclude: {
+                $null: true,
+              },
+            },
+            {
+              sitemap_exclude: {
+                $eq: false,
+              },
+            },
+          ],
+        },
+        populate: {
+          url_alias: {
+            populate: 'url_path',
+          },
+        },
+      },
     },
     orderBy: 'id',
     publicationState: excludeDrafts ? 'live' : 'preview',
@@ -228,8 +186,6 @@ const createSitemap = async (data) => {
 };
 
 export default () => ({
-  getFieldsFromConfig,
-  getRelationsFromConfig,
   getPages,
   getLocalizationIds,
   createSitemap,