+++ packages/core/admin/screens/Patterns/CreatePage/index.tsx
@@ -21,10 +21,9 @@ import {
   Loader,
 } from '@strapi/design-system';
 import { ArrowLeft, Check } from '@strapi/icons';
-import { request, useNotification } from '@strapi/helper-plugin';
-
+import { useNotification, useFetchClient } from '@strapi/helper-plugin';
 import schema from './utils/schema';
-
+import { ErrorResponse } from '../../../types/error-response';
 import pluginId from '../../../helpers/pluginId';
 import Center from '../../../components/Center';
 import Select from '../../../components/Select';
@@ -35,45 +34,41 @@ import { EnabledContentTypes } from '../../../types/enabled-contenttypes';
 import LanguageCheckboxes from '../../../components/LanguageCheckboxes';
 import HiddenLocalizedField from '../../../components/HiddenLocalizedField';
 
-const CreatePattternPage = () => {
+const CreatePatternPage = () => {
   const { push } = useHistory();
   const toggleNotification = useNotification();
   const [loading, setLoading] = useState(false);
   const [contentTypes, setContentTypes] = useState<EnabledContentTypes>([]);
   const { formatMessage } = useIntl();
+  const { get, post } = useFetchClient();
 
   useEffect(() => {
     setLoading(true);
-    request('/webtools/info/getContentTypes', { method: 'GET' })
-      .then((res: EnabledContentTypes) => {
-        setContentTypes(res);
+    get<EnabledContentTypes>('/webtools/info/getContentTypes')
+      .then((res) => {
+        const { data } = res;
+        setContentTypes(data);
         setLoading(false);
       })
       .catch(() => {
         setLoading(false);
       });
-  }, []);
+  }, [get]);
 
   const handleCreateSubmit = (
     values: PatternFormValues,
     { setSubmitting, setErrors }: FormikProps<PatternFormValues>,
   ) => {
-    request('/webtools/url-pattern/create', {
-      method: 'POST',
-      body: {
-        // @ts-ignore
-        data: values,
-      },
+    post('/webtools/url-pattern/create', {
+      data: values,
     })
       .then(() => {
         push(`/plugins/${pluginId}/patterns`);
         toggleNotification({ type: 'success', message: { id: 'webtools.settings.success.create' } });
         setSubmitting(false);
       })
-      .catch((err) => {
-        // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
-        if (err.response.payload[0].message === 'This attribute must be unique') {
-          // eslint-disable-next-line @typescript-eslint/no-unsafe-member-access
+      .catch((err: ErrorResponse) => {
+        if (err.response?.payload?.[0]?.message === 'This attribute must be unique') {
           setErrors({ code: err.response.payload[0].message as string });
         } else {
           toggleNotification({ type: 'warning', message: { id: 'notification.error' } });
@@ -85,21 +80,19 @@ const CreatePattternPage = () => {
   const validatePattern = async (values: PatternFormValues) => {
     const errors: FormikErrors<PatternFormValues> = {};
 
-    await request('/webtools/url-pattern/validate', {
-      method: 'POST',
-      body: {
-        // @ts-ignore
+    await post<ValidatePatternResponse>('/webtools/url-pattern/validate', {
+      data: {
         pattern: values.pattern,
         modelName: values.contenttype,
       },
     })
-      .then((res: ValidatePatternResponse) => {
-        if (res.valid === false) {
-          errors.pattern = res.message;
+      .then((res) => {
+        const response = res.data;
+        if (response.valid === false) {
+          errors.pattern = response.message;
         }
       })
-      .catch(() => {
-      });
+      .catch(() => {});
 
     return errors;
   };
@@ -131,12 +124,23 @@ const CreatePattternPage = () => {
       validate={validatePattern}
     >
       {({
-        handleSubmit, values, errors, touched, isSubmitting, setFieldValue,
+        handleSubmit,
+        values,
+        errors,
+        touched,
+        isSubmitting,
+        setFieldValue,
       }) => (
         <Form noValidate onSubmit={handleSubmit} placeholder={null}>
           <HeaderLayout
-            title={formatMessage({ id: 'webtools.settings.page.patterns.create.title', defaultMessage: 'Add new pattern' })}
-            subtitle={formatMessage({ id: 'webtools.settings.page.patterns.create.description', defaultMessage: 'Add a pattern for automatic URL alias generation.' })}
+            title={formatMessage({
+              id: 'webtools.settings.page.patterns.create.title',
+              defaultMessage: 'Add new pattern',
+            })}
+            subtitle={formatMessage({
+              id: 'webtools.settings.page.patterns.create.description',
+              defaultMessage: 'Add a pattern for automatic URL alias generation.',
+            })}
             as="h2"
             navigationAction={(
               <Link startIcon={<ArrowLeft />} to={`/plugins/${pluginId}/patterns`}>
@@ -145,7 +149,7 @@ const CreatePattternPage = () => {
                   defaultMessage: 'Back',
                 })}
               </Link>
-            )}
+                          )}
             primaryAction={(
               <Button type="submit" loading={isSubmitting} startIcon={<Check />}>
                 {formatMessage({
@@ -153,7 +157,7 @@ const CreatePattternPage = () => {
                   defaultMessage: 'Save',
                 })}
               </Button>
-            )}
+                          )}
           />
           <ContentLayout>
             <Stack spacing={7}>
@@ -185,9 +189,9 @@ const CreatePattternPage = () => {
                           defaultMessage: 'Content type',
                         })}
                         error={
-                          errors.contenttype && touched.contenttype
-                            ? formatMessage({ id: String(errors.contenttype), defaultMessage: 'Invalid value' })
-                            : null
+                            errors.contenttype && touched.contenttype
+                              ? formatMessage({ id: String(errors.contenttype), defaultMessage: 'Invalid value' })
+                              : null
                         }
                       />
                     </GridItem>
@@ -202,18 +206,18 @@ const CreatePattternPage = () => {
                     </GridItem>
                     <GridItem col={12} />
                     {(values.contenttype !== '') && (
-                      <GridItem col={6}>
-                        <PatternField
-                          values={values}
-                          uid={values.contenttype}
-                          setFieldValue={setFieldValue}
-                          error={
-                              errors.pattern && touched.pattern
-                                ? errors.pattern
-                                : null
-                            }
-                        />
-                      </GridItem>
+                    <GridItem col={6}>
+                      <PatternField
+                        values={values}
+                        uid={values.contenttype}
+                        setFieldValue={setFieldValue}
+                        error={
+                                  errors.pattern && touched.pattern
+                                    ? errors.pattern
+                                    : null
+                              }
+                      />
+                    </GridItem>
                     )}
                     <HiddenLocalizedField
                       localized={getSelectedContentType(values.contenttype)?.localized}
@@ -245,4 +249,4 @@ const CreatePattternPage = () => {
   );
 };
 
-export default CreatePattternPage;
+export default CreatePatternPage;